% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lux_to_relative.R
\name{lux_to_relative-shiny}
\alias{lux_to_relative-shiny}
\alias{lux_to_relative_output}
\alias{render_lux_to_relative}
\title{Shiny bindings for lux_to_relative}
\usage{
lux_to_relative_output(outputId, width = "100\%", height = "400px")

render_lux_to_relative(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a lux_to_relative}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using lux_to_relative within Shiny
applications and interactive Rmd documents.
}
