% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lux_to_relative.R
\name{lux_to_relative}
\alias{lux_to_relative}
\title{Relative time element}
\usage{
lux_to_relative(
  datetime,
  usetz = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{datetime}{A POSIXct scalar}

\item{usetz}{Whether to use the timezone or not}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}
}
\description{
Relative time element
}
