% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_time.R
\name{relative_time}
\alias{relative_time}
\title{Relative time}
\usage{
relative_time(datetime, usetz = FALSE)
}
\arguments{
\item{datetime}{A POSIXct scalar}

\item{usetz}{Whether to use the timezone or not}
}
\value{
A tag list element
}
\description{
Relative time
}
\examples{
relative_time(Sys.time())
}
